+++ recipes/debian.rb
@@ -23,6 +23,7 @@ if node['thruk']['version'].split('.')[0].to_i < 2
   remote_file "#{Chef::Config[:file_cache_path]}/#{node['thruk']['pkg_name']}" do
     action :create_if_missing
     source "#{node['thruk']['pkg_url']}/#{node['thruk']['pkg_name']}"
+    mode '0644'
     backup false
     not_if "dpkg-query -W thruk | grep -q '^thruk[[:blank:]]#{node['thruk']['version']}$'"
     notifies :install, 'dpkg_package[thruk]', :immediately
@@ -41,7 +42,7 @@ else
   apt_repository 'labs-thruk' do
     uri          "http://labs.consol.de/repo/stable/#{node['platform']}"
     key          'F8C1CA08A57B9ED7'
-    keyserver    'keys.gnupg.net'
+    keyserver    'hkp://keys.gnupg.net'
     distribution node['lsb']['codename']
     components   ['main']
   end